Extract text processing into new file
